% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_auth}
\alias{docu_auth}
\title{Create XML authentication string}
\usage{
docu_auth(username, password, integrator_key)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}
}
\description{
Create XML authentication string
}
