% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_embed}
\alias{docu_embed}
\title{Embedded docuSign}
\usage{
docu_embed(username = Sys.getenv("docuSign_username"),
  password = Sys.getenv("docuSign_password"),
  integrator_key = Sys.getenv("docuSign_integrator_key"), base_url,
  return_url, signer_name, client_user_id, uri)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{base_url}{docuSign baseURL}

\item{return_url}{URL to return to after signing}

\item{signer_name}{Name of person signing document}

\item{client_user_id}{ID for signer}

\item{uri}{docuSign uri}
}
\description{
Get URL for embedded docuSign
}
\examples{
# assuming env variables are properly set up
login <- docu_login()
env <- docu_envelope(
 account_id = login[1, "accountId"], base_url = login[1, "baseUrl"], 
 template_id = "e86ad42d-f935-4a95-8019-c9e2c902de15",
 template_roles = list(name = "Name", email = "email@example.com",
                      roleName = "Role"),
 email_subject = "Subject", email_blurb = "Body"
)
URL <- docu_embed(
 base_url = login[1, "baseUrl"], return_url = "www.google.com",
 signer_name = "Name", client_user_id = "1", 
 uri = env$uri
)
 
}
