% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_envelope}
\alias{docu_envelope}
\title{Create document for particular instance to be signed}
\usage{
docu_envelope(username = Sys.getenv("docuSign_username"),
  password = Sys.getenv("docuSign_password"),
  integrator_key = Sys.getenv("docuSign_integrator_key"), account_id,
  base_url, template_id, template_roles, email_subject, email_blurb)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{account_id}{docuSign accountId}

\item{base_url}{docuSign baseURL}

\item{template_id}{docuSign templateId}

\item{template_roles}{list of parameters passed to template}

\item{email_subject}{docuSign emailSubject}

\item{email_blurb}{docuSign emailBlurb}
}
\description{
Does envelope stuff
}
