% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_header}
\alias{docu_header}
\title{Create header for docuSign}
\usage{
docu_header(auth)
}
\arguments{
\item{auth}{XML object with authentication info}
}
\description{
Create header for authentication with docuSign
}
