% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_list_envelopes}
\alias{docu_list_envelopes}
\title{List envelopes since date}
\usage{
docu_list_envelopes(username = Sys.getenv("docuSign_username"),
  password = Sys.getenv("docuSign_password"),
  integrator_key = Sys.getenv("docuSign_integrator_key"), base_url, from_date)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{base_url}{base_url}

\item{from_date}{character indicating begin date of search}
}
\description{
List envelopes since date
}
\examples{
\dontrun{
login <- docu_login(demo = TRUE)
envelopes <- docu_list_envelopes(base_url = login$baseUrl[1], from_date = "2017/1/1")
}

}
