% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_login}
\alias{docu_login}
\title{Authenticate DocuSign}
\usage{
docu_login(username = Sys.getenv("docuSign_username"),
  password = Sys.getenv("docuSign_password"),
  integrator_key = Sys.getenv("docuSign_integrator_key"), demo = FALSE)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{demo}{indicate whether to use DocuSign demo server}
}
\description{
Login to DocuSign and get baseURL and accountId
}
\examples{
\dontrun{
# assuming env variables are properly set up
(login <- docu_login())
}
}
