% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_templates}
\alias{docu_templates}
\title{View templates}
\usage{
docu_templates(username = Sys.getenv("docuSign_username"),
  password = Sys.getenv("docuSign_password"),
  integrator_key = Sys.getenv("docuSign_integrator_key"), base_url)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{base_url}{docuSign baseURL}
}
\description{
See all templates associated with account
}
\examples{
\dontrun{
login <- docu_login()
templates <- docu_templates(base_url = login[1, 3])
}
}
