% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{parse_response}
\alias{parse_response}
\title{Process results from POST or GET}
\usage{
parse_response(response)
}
\arguments{
\item{response}{Result of POST or GET}
}
\description{
Process results from POST or GET
}
