% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_embed}
\alias{docu_embed}
\title{Embedded docuSign}
\usage{
docu_embed(username = Sys.getenv(), password = Sys.getenv(),
  integrator_key = Sys.getenv(), base_url, return_url, signer_name,
  client_user_id, uri)
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{base_url}{docuSign baseURL}

\item{return_url}{URL to return to after signing}

\item{signer_name}{Name of person signing document}

\item{client_user_id}{ID for signer}

\item{uri}{docuSign uri}
}
\description{
Get URL for embedded docuSign
}
