% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_embed}
\alias{docu_embed}
\title{Embedded docuSign}
\usage{
docu_embed(username = Sys.getenv("docuSign_username"),
  password = Sys.getenv("docuSign_password"),
  integrator_key = Sys.getenv("docuSign_integrator_key"), base_url,
  return_url, envelope_id, signer_name, signer_email, client_user_id,
  authentication_method = "None")
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}

\item{base_url}{docuSign baseURL}

\item{return_url}{URL to return to after signing}

\item{envelope_id}{ID for envelope returned from \code{docu_envelope}}

\item{signer_name}{Name of person signing document}

\item{signer_email}{Email of person signing document}

\item{client_user_id}{ID for signer}

\item{authentication_method}{Method application uses to authenticate user. Defaults to "None".}
}
\description{
Get URL for embedded docuSign
}
