% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{docu_login}
\alias{docu_login}
\title{Authenticate DocuSign}
\usage{
docu_login(username = Sys.getenv(), password = Sys.getenv(),
  integrator_key = Sys.getenv())
}
\arguments{
\item{username}{docuSign username}

\item{password}{docuSign password}

\item{integrator_key}{docusign integratorKey}
}
\description{
Login to DocuSign and get baseURL and accountId
}
